// src/app/rfq/models/inquiry.model.ts
export interface Inquiry {
  InqueryId: number;
  CategoryId: number;
  CategoryName: any[];
  CompanyName: any;
  FirstName: string;
  IncotermId: number;
  IncotermName: any;
  LastName: string;
  ProductNameById: any;
  ProductNameWithOutId: any;
  Quantity: number;
  UnitName: any;
  RFQId: number;
  Status: number;
  UnitId: number;
  UserId: number;
  VendorEmail: string;
  VendorName: any;
  UserName: any;
  VendorImagePath: string;
  ProductImagePath: string;
  UserImagePath: string;
}